+++ lib/src/importer/filesystem.dart
@@ -3,7 +3,7 @@
 // https://opensource.org/licenses/MIT.
 
 import '../importer.dart';
-import '../io.dart';
+import '../io.dart' as io;
 import '../util/path.dart';
 import 'result.dart';
 import 'utils.dart';
@@ -23,9 +23,11 @@ class FilesystemImporter extends Importer {
 
   ImporterResult load(Uri url) {
     var path = p.fromUri(url);
-    return new ImporterResult(readFile(path),
+    return new ImporterResult(io.readFile(path),
         sourceMapUrl: url, indented: p.extension(path) == '.sass');
   }
+
+  DateTime modificationTime(Uri url) => io.modificationTime(p.fromUri(url));
 
   String toString() => _loadPath;
 }